Added logic for encountering second variant of the disease to CA.
If using non-deterministic rules and there are 2 variants of the disease introduced, then when iterating, we can handle cases of both I and I' in the neighborhood, etc. and call appropriate variant map to give us percentProbability and if percent is not 0, we can transition the center cell states. For 0% prob, the centerCell stays as it is.