Fixed bug in iterateCABoard() for handling both variants of the disease
There was a bug with the logic of handling both I and I' in the neighborhood of the cell as well as either I and I' in the neighborhood. The solution was related with adding key to appropriate maps. I also added driver code in CA2d class to test a run/simulation of the CA until no I and I' on the board to make sure this works so I can start implementing GA.