Fixed small bugs in person, added epidemic spread after x time, etc.
Added infectionTillRecovery for a person no matter if they get quarantined or not, fixed bugs in person (setCurrentSIRQState() and peopleManager calling that was at fault) where a person was changing state after it recovered. Gui graph was not accounting for showing person who was quarantined and then recovered. Added an assumption.